requirements.txt
absl-py==0.7.0
asgiref==3.3.1
asn1crypto==0.24.0
astor==0.7.1
backports.shutil-get-terminal-size==1.0.0
backports.shutil-which==3.5.2
backports.weakref==1.0.post1
beautifulsoup4==4.6.3
cachetools==4.2.1
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
charset-normalizer==2.0.4
click==7.1.2
confusable-homoglyphs==3.2.0
cryptography==2.4.1
cssselect==1.1.0
cssutils==2.2.0
Django==2.0
django-appconf==1.0.4
django-bootstrap3==11.0.0
django-bootstrap4==0.0.7
django-braces==1.14.0
django-ckeditor==6.0.0
django-classy-tags==2.0.0
django-cors-headers==3.7.0
django-floppyforms==1.9.0
django-js-asset==1.2.2
django-map-widgets==0.3.0
django-model-utils==4.1.1
django-registration==3.1.1
django-restframework==0.0.1
django-star-ratings==0.9.1
djangorestframework==3.12.2
enum34==1.1.6
funcsigs==1.0.2
gast==0.2.2
geographiclib==1.50
geopy==2.1.0
gevent==1.4.0
greenlet==0.4.15
grpcio==1.18.0
h5py==2.8.0
icrawler==0.6.2
idna==2.7
importlib-metadata==3.7.3
ipaddress==1.0.22
iso-639==0.4.5
iso3166==0.9
isodate==0.6.0
jeepney==0.7.1
joblib==1.0.1
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
keyring==10.6.0
keyrings.alt==3.0
lxml==4.2.5
Markdown==3.0.1
mock==2.0.0
nltk==3.5
numpy==1.14.5
pbr==5.1.1
Pillow==5.4.1
premailer==3.7.0
protobuf==3.6.1
psycopg2==2.7.7
psycopg2-binary==2.7.5
pycparser==2.19
pycrypto==2.6.1
pycryptodome==3.7.0
pyOpenSSL==18.0.0
PySocks==1.6.8
pytz==2018.9
pyxdg==0.25
regex==2021.3.17
reportlab==3.5.66
requests==2.20.0
scapy==2.4.2
schedule==0.5.0
SecretStorage==2.3.1
singledispatch==3.4.0.3
six==1.12.0
SQLAlchemy==1.1.11
sqlparse==0.4.1
streamlink==0.14.2
swapper==1.1.2.post1
termcolor==1.1.0
tqdm==4.59.0
typing-extensions==3.7.4.3
urllib3==1.22
virtualenv==16.1.0
websocket-client==0.54.0
Werkzeug==0.14.1
yagmail==0.14.245
zipp==3.4.1
